+++ change-log.md
@@ -1,6 +1,7 @@
 TODO dsl实体查询加入以下（EntityQuery）
 
 1. fetch多次(大于1)，数据返回加入Object[] values(), <T extends Tuple> T tuple() 和 List<Object[]> listValues()，<T extends Tuple> List<T> listTuple() 
+2. 为mapper加入@Nonnull 支持，以及validation-api 支持再考虑
 
 # 0.7.4